Projects/Game Analysis/requirements.txt
pandas
matplotlib
beautifulsoup4
requests
numpy
pathlib
PyYAML